+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/MappingProfiles/LocationProfile.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using EasyMeets.Core.Common.DTO;
+using EasyMeets.Core.Common.DTO.Location;
 using EasyMeets.Core.DAL.Entities;
 
 namespace EasyMeets.Core.BLL.MappingProfiles
