+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Services/UserService.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using EasyMeets.Core.BLL.Interfaces;
-using EasyMeets.Core.Common.DTO;
+using EasyMeets.Core.Common.DTO.User;
 using EasyMeets.Core.DAL.Context;
 using Microsoft.EntityFrameworkCore;
 
@@ -10,7 +10,7 @@ namespace EasyMeets.Core.BLL.Services
     {
         public UserService(EasyMeetsCoreContext context, IMapper mapper) : base(context, mapper) {}
 
-        public async Task<UserDto> GetCurrentUserForAvailabilityPageAsync(int id)
+        public async Task<UserDto> GetCurrentUserAsync(int id)
         {
             var currentUser = await _context.Users.FirstOrDefaultAsync(x => x.Id == id);
             var currentUserDto = _mapper.Map<UserDto>(currentUser);
