+++ backend/EasyMeets.Core/EasyMeets.Core.Common/DTO/Availability/AvailabilitySlotDto.cs
@@ -1,5 +1,7 @@
 ﻿using EasyMeets.Core.Common.DTO.Common;
-using EasyMeets.Core.Common.Enums; 
+using EasyMeets.Core.Common.DTO.Location;
+using EasyMeets.Core.Common.DTO.User;
+using EasyMeets.Core.Common.Enums;
 
 namespace EasyMeets.Core.Common.DTO.Availability
 {
