+++ backend/EasyMeets.Core/EasyMeets.Core.Common/DTO/Location/LocationDto.cs
@@ -1,7 +1,7 @@
-﻿namespace EasyMeets.Core.Common.DTO
+﻿namespace EasyMeets.Core.Common.DTO.Location
 {
     public class LocationDto
-    { 
+    {
         public string Name { get; set; } = string.Empty;
     }
 }
