+++ backend/EasyMeets.Core/EasyMeets.Core.Common/DTO/User/UserDto.cs
@@ -1,14 +1,12 @@
 ﻿using EasyMeets.Core.Common.DTO.Common;
 
-namespace EasyMeets.Core.Common.DTO
+namespace EasyMeets.Core.Common.DTO.User
 {
     public class UserDto : EntityDto<long>
     {
         public string Name { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
         public string? PhoneNumber { get; set; }
-        public string? ImagePath { get; set; }
-        public string? Country { get; set; }
-        public string? TimeZone { get; set; }
+        public string? ImagePath { get; set; } 
     }
 }
