+++ backend/EasyMeets.Core/EasyMeets.Core.WebAPI/Controllers/AvailabilityController.cs
@@ -5,8 +5,6 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace EasyMeets.Core.WebAPI.Controllers
 {
-    //TODO: return Authorize
-    //[Authorize]
     [ApiController]
     [Route("[controller]")]
     public class AvailabilityController : ControllerBase
